Move Schemavalidation folder to DatabankLib
