--- Scripts/NMRlipidsDatabank.egg-info/PKG-INFO
@@ -1,14 +1,0 @@
-Metadata-Version: 2.4
-Name: NMRlipidsDatabank
-Version: 1.3.0
-Home-page: https://github.com/NMRlipids/Databank
-Author: NMRlipids open collaboration
-Author-email: samuli.ollila@helsinki.fi
-Requires-Python: >=3.9
-License-File: LICENSE.txt
-License-File: AUTHORS.md
-Dynamic: author
-Dynamic: author-email
-Dynamic: home-page
-Dynamic: license-file
-Dynamic: requires-python
