--- Scripts/NMRlipidsDatabank.egg-info/SOURCES.txt
@@ -1,29 +1,0 @@
-AUTHORS.md
-LICENSE.txt
-README.md
-pyproject.toml
-setup.py
-Scripts/DatabankLib/__init__.py
-Scripts/DatabankLib/analyze.py
-Scripts/DatabankLib/analyze_nmrpca.py
-Scripts/DatabankLib/core.py
-Scripts/DatabankLib/databankLibrary.py
-Scripts/DatabankLib/databankio.py
-Scripts/DatabankLib/databankop.py
-Scripts/DatabankLib/jpyroutines.py
-Scripts/DatabankLib/jsonEncoders.py
-Scripts/DatabankLib/maicos.py
-Scripts/DatabankLib/openmm_parser.py
-Scripts/DatabankLib/plottings.py
-Scripts/DatabankLib/quality.py
-Scripts/DatabankLib/utils.py
-Scripts/DatabankLib/settings/READMEexplanations.yaml
-Scripts/DatabankLib/settings/__init__.py
-Scripts/DatabankLib/settings/dic_lipids_UA.py
-Scripts/DatabankLib/settings/elements.py
-Scripts/DatabankLib/settings/engines.py
-Scripts/DatabankLib/settings/molecules.py
-Scripts/NMRlipidsDatabank.egg-info/PKG-INFO
-Scripts/NMRlipidsDatabank.egg-info/SOURCES.txt
-Scripts/NMRlipidsDatabank.egg-info/dependency_links.txt
-Scripts/NMRlipidsDatabank.egg-info/top_level.txt
