--- Scripts/NMRlipidsDatabank.egg-info/dependency_links.txt
@@ -1,1 +1,0 @@
-
